+++ adapters/vue/index.js
@@ -1,0 +1,1 @@
+export { useFoundry } from './useFoundry.js';